A `./wire` door is inert, so the projection gets its own
The three `wire.ts` files put a row's entry schema and the projection
that builds it in one module, and `surface.ts` imports the schema as a
VALUE — so the server's revision machinery (`@olai/surface/projection`,
and `@olai/format`'s set readers under it) landed on the graph of every
browser that loads the spec. A few hundred lines no tab can call, on the
one graph where nothing unused is free.

The split is by GRAPH rather than by subject: `./wire` is the schema and
crosses to the browser, `./projection` is the slice and does not.
