+++ packages/bundle/src/fence.test.ts
@@ -2003,7 +2003,7 @@ describe("only the registry knows a plugin's name in CODE, too", () => {
       "plugins/kolu/src/server.ts",
       "plugins/layout/src/Fault.tsx",
       "plugins/markdown/src/browser.tsx",
-      "plugins/markdown/src/wire.ts",
+      "plugins/markdown/src/projection.ts",
       "plugins/mcp/src/client.ts",
       "plugins/mcp/src/endpoint.ts",
       "plugins/navigation/src/browser.tsx",
@@ -2012,7 +2012,7 @@ describe("only the registry knows a plugin's name in CODE, too", () => {
       "plugins/outlines/src/browser/focus.ts",
       "plugins/outlines/src/browser/fold/memory.ts",
       "plugins/outlines/src/browser/page.ts",
-      "plugins/outlines/src/wire.ts",
+      "plugins/outlines/src/projection.ts",
       "plugins/search/src/table.bench.ts",
       "plugins/search/src/table.ts",
       "plugins/sidebar/src/Sidebar.tsx",
@@ -2026,7 +2026,7 @@ describe("only the registry knows a plugin's name in CODE, too", () => {
       "plugins/vault/src/browser/served.tsx",
       "plugins/vault/src/http/media.ts",
       "plugins/vault/src/lock.ts",
-      "plugins/vault/src/wire.ts",
+      "plugins/vault/src/projection.ts",
       "plugins/web-app/src/manifest.ts",
       "store/src/probe.ts",
       "store/src/store.ts",
