+++ packages/bundle/src/published.testlib.ts
@@ -84,9 +84,12 @@ import {
 import { seeded } from "@olai/format/testlib"
 import { decodedVault } from "@olai/format/testlib/scope"
 import type { Snapshot } from "@olai/store"
-import { type DocumentEntry, documentProjection } from "olai-plugin-markdown/wire"
-import { type OutlineEntry, outlineProjection } from "olai-plugin-outlines/wire"
-import { type Head, headProjection } from "olai-plugin-vault/wire"
+import type { DocumentEntry } from "olai-plugin-markdown/wire"
+import { documentProjection } from "olai-plugin-markdown/projection"
+import type { OutlineEntry } from "olai-plugin-outlines/wire"
+import { outlineProjection } from "olai-plugin-outlines/projection"
+import type { Head } from "olai-plugin-vault/wire"
+import { headProjection } from "olai-plugin-vault/projection"
 import { Result } from "effect"
 
 import type { Change } from "@olai/surface/projection"
