+++ packages/plugins/markdown/package.json
@@ -13,7 +13,8 @@
     "./surface": "./src/surface.ts",
     "./testids": "./src/testids.ts",
     "./client": "./src/client.ts",
-    "./wire": "./src/wire.ts"
+    "./wire": "./src/wire.ts",
+    "./projection": "./src/projection.ts"
   },
   "scripts": {
     "typecheck": "tsc --noEmit"
