+++ packages/plugins/markdown/src/server.ts
@@ -43,7 +43,8 @@ import { surface, faces, dispatch } from "./surface.ts"
 import { name } from "./name.ts"
 export { name } from "./name.ts"
 import type { Projection } from "@olai/surface/projection"
-import { documentProjection, type DocumentEntry } from "./wire.ts"
+import { documentProjection } from "./projection.ts"
+import type { DocumentEntry } from "./wire.ts"
 import type { FiledPageReading } from "@olai/format"
 import * as Bodies from "./server/bodies.ts"
