+++ packages/plugins/outlines/package.json
@@ -22,7 +22,8 @@
     "./slots": "./src/slots.ts",
     "./testids": "./src/testids.ts",
     "./client": "./src/client.ts",
-    "./wire": "./src/wire.ts"
+    "./wire": "./src/wire.ts",
+    "./projection": "./src/projection.ts"
   },
   "scripts": {
     "typecheck": "tsc --noEmit"
