+++ packages/plugins/outlines/src/server.ts
@@ -44,7 +44,8 @@ import { surface, faces, dispatch } from "./surface.ts"
 import { name } from "./name.ts"
 export { name } from "./name.ts"
 import type { Projection } from "@olai/surface/projection"
-import { outlineProjection, type OutlineEntry } from "./wire.ts"
+import { outlineProjection } from "./projection.ts"
+import type { OutlineEntry } from "./wire.ts"
 import { samePageReading, sameNarrowing, sameMoving } from "@olai/format"
 import type { FiledPageReading } from "@olai/format"
